feat: migrate /personInfo endpoint to AlertInfoController
- Migrated getPersonInfo endpoint from PersonController to AlertInfoController
- Moved business logic from PersonService to AlertInfoService
  and added @Autowired dependencies for PersonRepository and MedicalRecordRepository
- Refactored findByLastName in InMemoryPersonRepository and
  getMedicalRecordByFirstNameAndLastName in InMemoryMedicalRecordRepository with trim() for robustness
- Verified endpoint functionality at /personInfo?lastName=<lastName> using Postman
