+++ src/main/java/com/openclassrooms/safetynet/safetynetapi/controller/AlertInfoController.java
@@ -1,5 +1,6 @@
 package com.openclassrooms.safetynet.safetynetapi.controller;
 
+import com.openclassrooms.safetynet.safetynetapi.dto.PersonInfoDto;
 import com.openclassrooms.safetynet.safetynetapi.service.AlertInfoService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,4 +39,27 @@ public class AlertInfoController {
         }
     }
 
+
+    /**
+     * Handles GET requests to retrieve personal information filtered by last name.
+     *
+     * @param lastName the last name used to filter persons
+     * @return ResponseEntity containing:
+     * - HTTP 200 OK and a list of PersonInfoDto if matching persons are found,
+     * - HTTP 404 Not Found if no persons match the provided last name
+     */
+    @GetMapping("/personInfo")
+    public ResponseEntity<List<PersonInfoDto>> getPersonInfo(@RequestParam String lastName) {
+        log.info("Request received for /personInfo with lastName: {}", lastName);
+        List<PersonInfoDto> result = alertInfoService.getPersonInfoByLastName(lastName);
+
+        if (result.isEmpty()) {
+            log.warn("No persons found with lastName: {}", lastName);
+            return ResponseEntity.notFound().build();
+        }
+
+        log.info("{} person(s) found with lastName '{}'", result.size(), lastName);
+        return ResponseEntity.ok(result);
+
+    }
 }
