+++ src/main/java/com/openclassrooms/safetynet/safetynetapi/service/PersonService.java
@@ -3,7 +3,6 @@ package com.openclassrooms.safetynet.safetynetapi.service;
 import com.openclassrooms.safetynet.safetynetapi.dto.ChildDTO;
 import com.openclassrooms.safetynet.safetynetapi.dto.HouseholdMembersDTO;
 import com.openclassrooms.safetynet.safetynetapi.dto.PersonDTO;
-import com.openclassrooms.safetynet.safetynetapi.dto.PersonInfoDto;
 import com.openclassrooms.safetynet.safetynetapi.exception.PersonAlreadyExistsException;
 import com.openclassrooms.safetynet.safetynetapi.exception.PersonNotFoundException;
 import com.openclassrooms.safetynet.safetynetapi.model.MedicalRecord;
@@ -138,40 +137,6 @@ public class PersonService {
 
 
 
-    /**
-     * Retrieves a list of PersonInfoDto objects for all persons matching the given last name.
-     * <p>
-     * For each person found, this method fetches their medical record to include
-     * age (calculated from birthdate), medications, and allergies.
-     *
-     * @param lastName the last name to search for
-     * @return a list of PersonInfoDto containing personal and medical information
-     */
-    public List<PersonInfoDto> getPersonInfoByLastName(String lastName) {
-        List<Person> persons = personRepository.findByLastName(lastName);
-        List<PersonInfoDto> result = new ArrayList<>();
-
-        for (Person person : persons) {
-            MedicalRecord record = medicalRecordRepository.getMedicalRecordByFirstNameAndLastName(
-                    person.getFirstName(), person.getLastName());
-
-            if (record != null) {
-                int age = AgeUtil.calculateAge(record.getBirthdate());
-                PersonInfoDto dto = new PersonInfoDto(
-                        person.getFirstName(),
-                        person.getLastName(),
-                        person.getAddress(),
-                        person.getEmail(),
-                        age,
-                        record.getMedications(),
-                        record.getAllergies()
-                );
-                result.add(dto);
-            }
-        }
-
-        return result;
-    }
 
     /**
      * Retrieves a list of children (aged 18 or under) living at the specified address,
